[Refactor] Extract WebSocketBufferOverflowTest
