+++ async-http-client-backend/src/test/scala/sttp/client/asynchttpclient/AsyncHttpClientWebSocketTest.scala
@@ -1,37 +1,13 @@
 package sttp.client.asynchttpclient
 
-import java.nio.channels.ClosedChannelException
+import sttp.client.testing.websocket.{WebSocketBufferOverflowTest, WebSocketStreamingTest, WebSocketTest}
 
-import sttp.client.{asWebSocketAlways, basicRequest}
-import sttp.client.monad.syntax._
-import sttp.client.testing.websocket.WebSocketTest
-import sttp.model.Uri._
-import sttp.client.testing.HttpTest.wsEndpoint
-import sttp.client.testing.WebSocketStreamingTest
-import sttp.client.ws.WebSocket
-
-import scala.concurrent.duration._
-
-abstract class AsyncHttpClientWebSocketTest[F[_], S] extends WebSocketTest[F] with WebSocketStreamingTest[F, S] {
-
-  it should "error if incoming messages overflow the buffer" in {
-    basicRequest
-      .get(uri"$wsEndpoint/ws/echo")
-      .response(asWebSocketAlways { ws: WebSocket[F] =>
-        send(ws, AsyncHttpClientBackend.DefaultWebSocketBufferCapacity.get + 1) >>
-          eventually(10.millis, 500) {
-            ws.isOpen.map(_ shouldBe false)
-          }
-      })
-      .send()
-      .map(_.body)
-      .handleError {
-        case _: ClosedChannelException => succeed.unit
-      }
-      .toFuture()
-  }
+abstract class AsyncHttpClientWebSocketTest[F[_], S]
+    extends WebSocketTest[F]
+    with WebSocketStreamingTest[F, S]
+    with WebSocketBufferOverflowTest[F] {
 
   override def throwsWhenNotAWebSocket: Boolean = true
 
-  def eventually[T](interval: FiniteDuration, attempts: Int)(f: => F[T]): F[T]
+  override def bufferCapacity: Int = AsyncHttpClientBackend.DefaultWebSocketBufferCapacity.get
 }
